Switch format for nuer
