+++ administration/src/project-configs/bayern/config.ts
@@ -1,10 +1,12 @@
 import { ProjectConfig } from '../getProjectConfig'
 import { dataPrivacyBaseHeadline, DataPrivacyBaseText } from './dataPrivacyBase'
+import { createEmptyBavariaCard } from '../../cards/cardBlueprints'
 
 const config: ProjectConfig = {
   name: 'Ehrenamtskarte Bayern',
   projectId: 'bayern.ehrenamtskarte.app',
   applicationFeatureEnabled: true,
+  createEmptyCard: createEmptyBavariaCard,
   dataPrivacyHeadline: dataPrivacyBaseHeadline,
   dataPrivacyContent: DataPrivacyBaseText,
 }
