+++ administration/src/project-configs/nuernberg/config.ts
@@ -1,10 +1,12 @@
 import { ProjectConfig } from '../getProjectConfig'
+import { createEmptyNuernbergCard } from '../../cards/cardBlueprints'
 import { dataPrivacyBaseHeadline, DataPrivacyBaseText } from './dataPrivacyBase'
 
 const config: ProjectConfig = {
   name: 'Digitaler Nürnberg-Pass',
   projectId: 'nuernberg.sozialpass.app',
   applicationFeatureEnabled: false,
+  createEmptyCard: createEmptyNuernbergCard,
   dataPrivacyHeadline: dataPrivacyBaseHeadline,
   dataPrivacyContent: DataPrivacyBaseText,
 }
