+++ administration/src/project-configs/showcase/config.ts
@@ -1,10 +1,12 @@
 import { ProjectConfig } from '../getProjectConfig'
+import { createEmptyBavariaCard } from '../../cards/cardBlueprints'
 import { dataPrivacyBaseHeadline, DataPrivacyBaseText } from '../bayern/dataPrivacyBase'
 
 const config: ProjectConfig = {
   name: 'Showcase Berechtigungskarte',
   projectId: 'showcase.entitlementcard.app',
   applicationFeatureEnabled: true,
+  createEmptyCard: createEmptyBavariaCard,
   dataPrivacyHeadline: dataPrivacyBaseHeadline,
   dataPrivacyContent: DataPrivacyBaseText,
 }
